fix: 체결 순환참조 제거 (추후 리팩토링 필요)
